moved C shared lib go wrapper in indepdnent driectory
