+++ .github/workflows/build_shared.yml
@@ -58,19 +58,19 @@ jobs:
           go-version: '1.21'
 
       - name: Build shared library for Linux
-        working-directory: python
+        working-directory: sharedlib
         run: |
-          go build -buildmode=c-shared -o package/libprismaid_linux_amd64.so export.go
+          go build -buildmode=c-shared -o libprismaid_linux_amd64.so export.go
 
       - name: Upload Linux shared library artifact
         uses: actions/upload-artifact@v4
         with:
           name: linux-shared-library
-          path: python/package/libprismaid_linux_amd64.so
+          path: sharedlib/libprismaid_linux_amd64.so
 
       - name: Commit and push changes
         run: |
-          cp python/package/libprismaid_linux_amd64.so r/prismaid/inst/libs/libprismaid_linux_amd64.so
+          cp sharedlib/libprismaid_linux_amd64.so r/prismaid/inst/libs/libprismaid_linux_amd64.so
           git config --local user.email "action@github.com"
           git config --local user.name "GitHub Action"
           git pull --no-rebase origin ${GITHUB_REF#refs/heads/}
@@ -93,20 +93,20 @@ jobs:
           go-version: '1.21'
 
       - name: Build shared library for Windows
-        working-directory: python
+        working-directory: sharedlib
         shell: pwsh
         run: |
-          go build -buildmode=c-shared -o package/libprismaid_windows_amd64.dll .\export.go
+          go build -buildmode=c-shared -o libprismaid_windows_amd64.dll .\export.go
 
       - name: Upload Windows shared library artifact
         uses: actions/upload-artifact@v4
         with:
           name: windows-shared-library
-          path: python/package/libprismaid_windows_amd64.dll
+          path: sharedlib/libprismaid_windows_amd64.dll
 
       - name: Commit and push changes
         run: |
-          cp python/package/libprismaid_windows_amd64.dll r/prismaid/inst/libs/libprismaid_windows_amd64.dll
+          cp sharedlib/libprismaid_windows_amd64.dll r/prismaid/inst/libs/libprismaid_windows_amd64.dll
           git config --local user.email "action@github.com"
           git config --local user.name "GitHub Action"
           git pull --no-rebase origin ${GITHUB_REF#refs/heads/}
@@ -129,19 +129,19 @@ jobs:
           go-version: '1.21'
 
       - name: Build shared library for macOS
-        working-directory: python
+        working-directory: sharedlib
         run: |
-          go build -buildmode=c-shared -o package/libprismaid_darwin_amd64.dylib export.go
+          go build -buildmode=c-shared -o libprismaid_darwin_amd64.dylib export.go
 
       - name: Upload macOS shared library artifact
         uses: actions/upload-artifact@v4
         with:
           name: macos-shared-library
-          path: python/package/libprismaid_darwin_amd64.dylib
+          path: sharedlib/libprismaid_darwin_amd64.dylib
 
       - name: Commit and push changes
         run: |
-          cp python/package/libprismaid_darwin_amd64.dylib r/prismaid/inst/libs/libprismaid_darwin_amd64.dylib
+          cp sharedlib/libprismaid_darwin_amd64.dylib r/prismaid/inst/libs/libprismaid_darwin_amd64.dylib
           git config --local user.email "action@github.com"
           git config --local user.name "GitHub Action"
           git pull --no-rebase origin ${GITHUB_REF#refs/heads/}
